Rename type/ to definitions/
