+++ include/definitions/drawn_item_t.hpp
