+++ include/extensions/iostream_extensions.hpp
@@ -2,7 +2,7 @@
 
 #include <iostream>
 
-#include "type/coordinate_t.hpp"
+#include "definitions/coordinate_t.hpp"
 #include "s2.hpp"
 
 namespace ingress_drone_explorer {
