+++ src/s2.cpp
@@ -3,7 +3,7 @@
 #include <algorithm>
 #include <cmath>
 
-#include "type/coordinate_t.hpp"
+#include "definitions/coordinate_t.hpp"
 
 namespace ingress_drone_explorer {
